Add site verification, tests; modularize validators
